Implement Channel 2 (flashing LEDs)
Required a HUGE rewrite of how I handle setting LEDs normally, because every flashing action had to be cleared when an LED was normally set. hal_plot_led was mostly replaced with rgb_led or novation_led everywhere in the firmware to accomodate for this.

Also, don't update LEDs over SysEx if we're in a special mode (mode < 128 must apply).